# api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/social/connections": {
      "get": {
        "operationId": "SocialConnectionsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocialConnectionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token thiếu, sai hoặc hết hạn"
          },
          "403": {
            "description": "Không phải tài khoản creator"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Danh sách tài khoản mạng xã hội đã kết nối",
        "tags": [
          "Social-Connections"
        ]
      }
    },
    "/social/{platform}/authorize": {
      "get": {
        "operationId": "SocialConnectionsController_authorize",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "tiktok",
                "instagram",
                "youtube",
                "facebook"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "authorizationUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token thiếu, sai hoặc hết hạn"
          },
          "403": {
            "description": "Không phải tài khoản creator"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Lấy URL chuyển hướng để bắt đầu luồng OAuth",
        "tags": [
          "Social-Connections"
        ]
      }
    },
    "/social/{platform}/callback": {
      "post": {
        "operationId": "SocialConnectionsController_callback",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "tiktok",
                "instagram",
                "youtube",
                "facebook"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialOAuthCallbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialConnectionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "State hết hạn, sai, hoặc token hỏng"
          },
          "403": {
            "description": "State thuộc về creator khác, hoặc không phải creator"
          },
          "409": {
            "description": "Tài khoản mạng xã hội này đã thuộc creator khác"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Hoàn tất luồng OAuth và lưu kết nối",
        "tags": [
          "Social-Connections"
        ]
      }
    },
    "/social/connections/{platform}": {
      "delete": {
        "operationId": "SocialConnectionsController_disconnect",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "tiktok",
                "instagram",
                "youtube",
                "facebook"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Đã ngắt kết nối"
          },
          "401": {
            "description": "Token thiếu, sai hoặc hết hạn"
          },
          "403": {
            "description": "Không phải tài khoản creator"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Ngắt kết nối một nền tảng",
        "tags": [
          "Social-Connections"
        ]
      }
    }
  },
  "info": {
    "title": "CRM KOC API",
    "description": "REST API for the CRM KOC system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ESocialPlatform": {
        "type": "string",
        "enum": [
          "tiktok",
          "instagram",
          "youtube",
          "facebook"
        ]
      },
      "SocialConnectionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "platform": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ESocialPlatform"
              }
            ]
          },
          "externalAccountId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "biography": {
            "type": "string",
            "nullable": true
          },
          "profileUrl": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "followerCount": {
            "type": "string",
            "description": "bigint nên driver trả về chuỗi"
          },
          "followingCount": {
            "type": "string"
          },
          "contentCount": {
            "type": "string"
          },
          "grantedScopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tokenExpiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "lastSyncedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "syncError": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform",
          "externalAccountId",
          "username",
          "displayName",
          "biography",
          "profileUrl",
          "avatarUrl",
          "followerCount",
          "followingCount",
          "contentCount",
          "grantedScopes",
          "tokenExpiresAt",
          "lastSyncedAt",
          "syncError",
          "isActive",
          "createdAt"
        ]
      },
      "SocialOAuthCallbackDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Authorization code do nền tảng trả về"
          },
          "state": {
            "type": "string",
            "description": "Giá trị state của lần chuyển hướng đi"
          }
        },
        "required": [
          "code",
          "state"
        ]
      }
    }
  }
}